Experemental stream compressor script
A rough stream compression script is made to begin expereminting
with program structure and compression techinques. A simple version
of the Golomb-Rice encoder is implemented which codes first order
difference accelerometer information. Compression rates are around
93% and 67% for csv file and packed data respectively. Compare this
to tar performance on the csv of 90%. Note that only one data file
was used for this measurement and kinematics were mostly stationary
during the period. I expect around 60% compression of packed data on
average for all kinematic activity.